requirements.txt
# Core
typer==0.16.0
rich==13.7.1
python-dotenv

# LLM и генерация
litellm
openai

# Git и версия навыков
GitPython==3.1.43

# База данных
SQLAlchemy==2.0.32
sqlite-utils==3.36

# Тесты
PyYAML==6.0.2
pytest==8.3.2

# Runtime
watchdog==4.0.1

# Audio / Voice
sounddevice==0.4.7        # Аудио интерфейс для микрофона и воспроизведения
vosk==0.3.45              # Локальный ASR (Speech-to-Text)
pvporcupine==2.2.0        # Wake Word Detection (альтернатива snowboy)

# MQTT (понадобится позже)
paho-mqtt==2.1.0
